fixed a problem with asdouble

+++ src/ch/epfl/javelo/Q28_4.java
@@ -19,7 +19,7 @@ public final class Q28_4 {
      */
 
     public static double asDouble(int q28_4){
-        return Math.scalb(q28_4, -4);
+        return Math.scalb((double) q28_4, -4);
     }
 
     /**
